+++ admin-portal/src/pages/aidPackage/aidPackage.tsx
@@ -1,11 +1,13 @@
 import React, { useEffect, useState } from "react";
-import { useAsyncDebounce } from "react-table";
 import toast from "react-simple-toasts";
 import { Stepper, Step } from "../../components/stepper";
 import MedicalNeedsService from "../../apis/services/MedicalNeedsService";
 import AidPackageService from "../../apis/services/AidPackageService";
-import { AidPackage } from "../../types/AidPackage";
-import { Quotation } from "../../types/Quotation";
+import {
+  AidPackage,
+  DraftAidPackages,
+  NeedAssignments,
+} from "../../types/AidPackage";
 import { MedicalNeed } from "../../types/MedicalNeeds";
 import ManageAidPackages from "./manageAidPackages/manageAidPackages";
 import AssignSuppliers from "./assignSuppliers/assignSuppliers";
@@ -15,23 +17,6 @@ enum STEPS {
   ASSIGN_SUPPLIERS,
   MANAGE_AID_PACKAGES,
 }
-
-export type NeedAssignment = Map<number, number | null>; // Map<supplierId: quantity>
-
-export type NeedAssignments = {
-  [needID: string]: NeedAssignment;
-};
-
-export type DraftAidPackage = {
-  supplierID: number;
-  period: Quotation["period"];
-  name: string;
-  details: string;
-  isPublished?: boolean;
-};
-export type DraftAidPackages = {
-  [key: string]: DraftAidPackage;
-};
 
 export default function CreateAidPackage() {
   const [currentFormStep, setCurrentFormStep] = useState(0);
@@ -185,44 +170,6 @@ export default function CreateAidPackage() {
             Next
           </button>
         )}
-      </div>
-    </div>
-  );
-}
-
-export function GlobalFilter({
-  globalFilter,
-  setGlobalFilter,
-}: {
-  globalFilter: any;
-  setGlobalFilter: any;
-}) {
-  const [value, setValue] = useState(globalFilter);
-  const onChange = useAsyncDebounce((filter) => {
-    setGlobalFilter(filter || undefined);
-  }, 50);
-
-  return (
-    <div
-      className="packageTableSearch"
-      style={{
-        position: "sticky",
-        top: 0,
-        background: "white",
-        paddingBottom: "1rem",
-      }}
-    >
-      <div className="searchContainer">
-        <img src="/assets/svg/search_icon.svg" alt="search-icon" />
-        <input
-          placeholder="Search"
-          className="textField"
-          value={value || ""}
-          onChange={(e) => {
-            setValue(e.target.value);
-            onChange(e.target.value);
-          }}
-        />
       </div>
     </div>
   );
